tests/requirements.txt
appium-python-client>=3.1.0
selenium>=4.18.0
pytest>=7.4.0
PyYAML>=6.0.1
